+++ lib/src/platform/plaid_method_channel.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:flutter/services.dart';
 
 import '../core/events.dart';
@@ -6,12 +8,10 @@ import 'plaid_platform_interface.dart';
 
 class PlaidMethodChannel extends PlaidPlatformInterface {
   /// The method channel used to interact with the native platform.
-  final MethodChannel _channel =
-      const MethodChannel('plugins.flutter.io/plaid_flutter');
+  final MethodChannel _channel = const MethodChannel('plugins.flutter.io/plaid_flutter');
 
   /// The event channel used to receive changes from the native platform.
-  final EventChannel _eventChannel =
-      const EventChannel('plugins.flutter.io/plaid_flutter/events');
+  final EventChannel _eventChannel = const EventChannel('plugins.flutter.io/plaid_flutter/events');
 
   /// A broadcast stream from the native platform
   Stream<LinkObject>? _onObject;
@@ -63,5 +63,21 @@ class PlaidMethodChannel extends PlaidPlatformInterface {
   @override
   Future<void> submit(SubmissionData data) async {
     await _channel.invokeMethod('submit', data.toJson());
+  }
+
+  /// It allows the client application to submit additional user-collected data to the Link flow (e.g. a user phone number) for the Layer product.
+  @override
+  Future<void> syncFinanceKit(String token, bool requestAuthorizationIfNeeded) async {
+    if (Platform.isIOS) {
+      await _channel.invokeMethod(
+        'syncFinanceKit',
+        {
+          "token": token,
+          "requestAuthorizationIfNeeded": requestAuthorizationIfNeeded,
+        },
+      );
+    } else {
+      throw UnimplementedError('syncFinanceKit is only implemented for iOS.');
+    }
   }
 }